% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{dof.hypothesis_test}
\alias{dof.hypothesis_test}
\title{Degrees of freedom method for hypothesis tests}
\usage{
\method{dof}{hypothesis_test}(x, ...)
}
\arguments{
\item{x}{a hypothesis test}

\item{...}{additional arguments}
}
\value{
degrees of freedom
}
\description{
Degrees of freedom method for hypothesis tests
}
