% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{dof}
\alias{dof}
\alias{dof.hypothesis_test}
\title{Extract the degrees of freedom from a hypothesis test}
\usage{
dof(x, ...)

\method{dof}{hypothesis_test}(x, ...)
}
\arguments{
\item{x}{a hypothesis test object}

\item{...}{additional arguments to pass into the method}
}
\value{
Numeric degrees of freedom.
}
\description{
Extract the degrees of freedom from a hypothesis test
}
\examples{
w <- wald_test(estimate = 2.5, se = 0.8)
dof(w)
}
