% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\docType{package}
\name{hypothesize-package}
\alias{hypothesize}
\alias{hypothesize-package}
\title{hypothesize: Consistent API for Hypothesis Testing}
\description{
Provides a consistent API for hypothesis testing built on principles from 'Structure and Interpretation of Computer Programs': data abstraction, closure (combining tests yields tests), and higher-order functions (transforming tests). Implements z-tests, Wald tests, likelihood ratio tests, Fisher's method for combining p-values, and multiple testing corrections. Designed for use by other packages that want to wrap their hypothesis tests in a consistent interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/queelius/hypothesize}
  \item \url{https://queelius.github.io/hypothesize/}
  \item Report bugs at \url{https://github.com/queelius/hypothesize/issues}
}

}
\author{
\strong{Maintainer}: Alexander Towell \email{lex@metafunctor.com} (\href{https://orcid.org/0000-0001-6443-9897}{ORCID})

}
\keyword{internal}
