% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{is_significant_at.hypothesis_test}
\alias{is_significant_at.hypothesis_test}
\title{Significance test for the hypothesis_test class.}
\usage{
\method{is_significant_at}{hypothesis_test}(x, alpha, ...)
}
\arguments{
\item{x}{a hypothesis test object}

\item{alpha}{significance level}
}
\value{
logical indicating whether the test is significant at the given
significance level alpha (default is 0.05) or not.
}
\description{
Significance test for the hypothesis_test class.
}
