% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{pval.hypothesis_test}
\alias{pval.hypothesis_test}
\title{p-value method for hypothesis tests}
\usage{
\method{pval}{hypothesis_test}(x, ...)
}
\arguments{
\item{x}{a hypothesis test}

\item{...}{additional arguments}
}
\value{
p-value
}
\description{
p-value method for hypothesis tests
}
