% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{pval}
\alias{pval}
\alias{pval.hypothesis_test}
\title{Extract the p-value from a hypothesis test}
\usage{
pval(x, ...)

\method{pval}{hypothesis_test}(x, ...)
}
\arguments{
\item{x}{a hypothesis test object}

\item{...}{additional arguments to pass into the method}
}
\value{
Numeric p-value.
}
\description{
Extract the p-value from a hypothesis test
}
\examples{
w <- wald_test(estimate = 2.5, se = 0.8)
pval(w)
}
