% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{test_stat.hypothesis_test}
\alias{test_stat.hypothesis_test}
\title{Test statistic method for hypothesis tests}
\usage{
\method{test_stat}{hypothesis_test}(x, ...)
}
\arguments{
\item{x}{a hypothesis test}

\item{...}{additional arguments}
}
\value{
test statistic
}
\description{
Test statistic method for hypothesis tests
}
