% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{test_stat}
\alias{test_stat}
\alias{test_stat.hypothesis_test}
\title{Extract the test statistic from a hypothesis test}
\usage{
test_stat(x, ...)

\method{test_stat}{hypothesis_test}(x, ...)
}
\arguments{
\item{x}{a hypothesis test object}

\item{...}{additional arguments to pass into the method}
}
\value{
Numeric test statistic.
}
\description{
Extract the test statistic from a hypothesis test
}
\examples{
w <- wald_test(estimate = 2.5, se = 0.8)
test_stat(w)
}
