% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{upper}
\alias{upper}
\title{Extract the upper bound of a confidence set}
\usage{
upper(x, ...)
}
\arguments{
\item{x}{a confidence_set object}

\item{...}{additional arguments (ignored)}
}
\value{
numeric upper bound
}
\description{
Extract the upper bound of a confidence set
}
