% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{dof}
\alias{dof}
\title{Generic method for extracting the degrees of freedom from a hypothesis test}
\usage{
dof(x, ...)
}
\arguments{
\item{x}{a hypothesis test object}

\item{...}{additional arguments to pass into the method}
}
\value{
degrees of freedom
}
\description{
Generic method for extracting the degrees of freedom from a hypothesis test
}
