% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{is_significant_at}
\alias{is_significant_at}
\title{Generic method for checking if a hypothesis test is significant at a given
significance level.}
\usage{
is_significant_at(x, alpha, ...)
}
\arguments{
\item{x}{a hypothesis test object}

\item{alpha}{significance level}

\item{...}{additional arguments passed to methods}
}
\value{
logical indicating whether the test is significant at the given
significance level alpha (e.g., 0.05) or not.
}
\description{
Generic method for checking if a hypothesis test is significant at a given
significance level.
}
