% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{lower}
\alias{lower}
\title{Extract the lower bound of a confidence set}
\usage{
lower(x, ...)
}
\arguments{
\item{x}{a confidence_set object}

\item{...}{additional arguments (ignored)}
}
\value{
numeric lower bound
}
\description{
Extract the lower bound of a confidence set
}
