% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{lrt}
\alias{lrt}
\title{Likelihood ratio test. This function computes the likelihood ratio test
statistic and p-value.}
\usage{
lrt(null_loglik, alt_loglik, dof)
}
\arguments{
\item{null_loglik}{the likelihood value from a simpler likelihood model nested
within the alternative model}

\item{alt_loglik}{the likelihood value from the more complicated model}

\item{dof}{degrees of freedom}
}
\value{
likelihood ratio test
}
\description{
Assumptions:
(1) The null model is nested within the alternative model.
(2) The likelihood values are computed from the same data set.
}
\details{
The degrees of freedom are conceptualized in the context of vector spaces.
Thus, by (1), the parameter space of the null model is a subspace of the
parameter space of the alternative model, both of which are vector spaces.
The degrees of freedom is the difference in the dimension of the null model
and the alternative model. Normally, this boils down to the difference in
the number of parameters between the two models, e.g., fixing some of the
components in the null model to some constant value.

The likelihood ratio test statistic is defined as:

LRT = -2 (null_loglik - alt_loglik)

which, under the null hypothesis, is asymptotically chi-squared distributed
with degrees of freedom equal to the dof previously defined. Thus, we can
ask whether the observed likelihood ratio test statistic is significantly
different from the expected value under the null hypothesis. The p-value is
the probability of observing an LRT statistic as extreme as the one observed
under the null hypothesis.
}
\examples{
# create a likelihood ratio test
lrt <- lrt(null_loglik = -100, alt_loglik = -90, dof = 1)
# print the test
test_stat(lrt)
pval(lrt)
dof(lrt)
# check if the test is significant at the 5\% level
is_significant_at(lrt, 0.05)

}
