% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{print.confidence_set}
\alias{print.confidence_set}
\title{Print method for confidence sets}
\usage{
\method{print}{confidence_set}(x, ...)
}
\arguments{
\item{x}{a confidence_set object}

\item{...}{additional arguments (ignored)}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print method for confidence sets
}
