% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{print.hypothesis_test}
\alias{print.hypothesis_test}
\title{Print method for hypothesis tests}
\usage{
\method{print}{hypothesis_test}(x, ...)
}
\arguments{
\item{x}{a hypothesis test}

\item{...}{additional arguments}
}
\value{
a string representation of the hypothesis test
}
\description{
Print method for hypothesis tests
}
