% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{pval}
\alias{pval}
\title{Generic method for extracting the p-value from a hypothesis test}
\usage{
pval(x, ...)
}
\arguments{
\item{x}{a hypothesis test object}

\item{...}{additional arguments to pass into the method}
}
\value{
p-value
}
\description{
Generic method for extracting the p-value from a hypothesis test
}
