% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{test_stat}
\alias{test_stat}
\title{Generic method for extracting the test statistic from a hypothesis test}
\usage{
test_stat(x, ...)
}
\arguments{
\item{x}{a hypothesis test object}

\item{...}{additional arguments to pass into the method}
}
\value{
test statistic
}
\description{
Generic method for extracting the test statistic from a hypothesis test
}
